fix(multiclass): widen maxClasses to Long for JS-sentinel decoding (#1143)
## Summary
- Production crash-looped on boot because `application-local.yaml`
carried `maxClasses: 9007199254740991` (`Number.MAX_SAFE_INTEGER`, the
JS "unlimited" sentinel from our config-authoring tooling) and Hoplite
refused to decode that into `MulticlassConfig.maxClasses: Int`.
- Widen `MulticlassConfig.maxClasses` and the matching `Trainer.List`
GMCP field to `Long`. Wire format is unchanged (TS already uses
`number`; the sentinel fits exactly in `Number`).
- Added a regression test in `AppConfigLoaderTest` that decodes the
sentinel via Hoplite end-to-end.

## Test plan
- [x] `./gradlew ktlintCheck` — passes
- [x] `./gradlew test --tests "dev.ambon.config.AppConfigLoaderTest"
--tests "dev.ambon.engine.commands.TrainerUnlockEconomicsTest"` — passes
(covers the Hoplite path + trainer comparison + GMCP DTO emission)
- [ ] Once merged, deploy to prod and confirm `application-local.yaml`
no longer crashes the loader (the original stack trace was the `Could
not decode 9007199254740991 into a Int` at
`application-local.yaml:2084:18`).

+++ src/main/kotlin/dev/ambon/engine/GmcpEmitter.kt
@@ -2460,7 +2460,7 @@ class GmcpEmitter(
         /** Gold cost for *this player's next* unlock; already includes the exponential scaling. */
         val multiclassGoldCost: Long,
         /** Hard cap on the player's total unlocked classes (including starter). */
-        val multiclassMaxClasses: Int,
+        val multiclassMaxClasses: Long,
         /** Player's current unlocked-class count (including starter). */
         val multiclassUnlockedCount: Int,
         /**
